과제 리팩터링
