+++ src/hanmin/StudentDetail.java
@@ -1,4 +1,4 @@
-package Hanmin;
+package hanmin;
 
 import java.util.List;
 import java.util.Optional;
